requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django>=4.0,<5.0
djangorestframework
gunicorn
psycopg2-binary
drf-yasg